Re-connect for each job
Keeping a long lived connection was very flaky - after couple of hours
it would always fail.

It would make sense to use cron but that would be complicated because of credentials (we wanna keep it simple but secure). It wouldn't be possible to only keep the credentials in memory. One solution could be having a separate long-lived process (keychain) keeping the credentials and the cronjob only asking for it. I'm however not sure if inter-process communication is secure, sounds like that could be intercepted if somebody gets access to the machine.